the rijiggening  🐿 v2.5.16
